+++ it/proxy_test.py
@@ -23,9 +23,6 @@ import pytest
 import it
 from esrally.utils import process
 
-# pylint: disable=unused-import
-from it import fresh_log_file
-
 HttpProxy = collections.namedtuple("HttpProxy", ["authenticated_url", "anonymous_url"])
